minor bugfixes

+++ Practicum/Week 04 Constructors/Task03/time.h
@@ -12,7 +12,7 @@ class Time
 {
     double time;
 public:
-    Time(double hours, double minutes);
+    Time(int hours, int minutes);
 
     
 };
